✅ test: split vote-reason export test into sibling file (#609)
The added rendersVoteReasonAsNestedThought pushed
ResultMarkdownExporterTests.swift over the 400-line file_length cap. Move
it to a sibling extension file (not a new @Suite, per .claude/rules/testing.md)
sharing the suite's makeExporter/makeTurn helpers.
